+++ src/providers/imgix/runtime.ts
@@ -1,5 +1,6 @@
 import { RuntimeProvider, ImageModifiers } from 'types'
-import { cleanDoubleSlashes, createOperationsGenerator } from '../../runtime/provider-utils'
+import { cleanDoubleSlashes, createOperationsGenerator } from '~image/utils'
+import fetch from '~image/fetch'
 
 const operationsGenerator = createOperationsGenerator({
   keyMap: {
