+++ src/runtime/fetch.ts
@@ -1,0 +1,5 @@
+import { cleanDoubleSlashes } from './utils'
+
+export default function imageFetch (url: string) {
+  return fetch(cleanDoubleSlashes(url))
+}
